Fix new lines, trans emoji

+++ stickers.py
@@ -2,10 +2,15 @@ import os
 import hashlib
 
 def deleteSticker(sticker):
-  os.unlink(sticker)
+  if os.path.exists(sticker):
+    os.unlink(sticker)
 
 def validSize(sticker):
-  return os.path.getsize(sticker) < 64000
+  if os.path.exists(sticker):
+    return os.path.getsize(sticker) < 64000
+  else:
+    print("ERROR no sticker could be exported")
+    return False
 
 def tempPath(input):
   return './temp/' + hashlib.md5(input.encode('utf-8')).hexdigest() + '.tgs'
